+++ custom_components/owlet/__init__.py
@@ -10,8 +10,6 @@ import asyncio
 import logging
 
 from aiohttp import ClientError
-from pyowletapi.api import OwletAPI
-from pyowletapi.sock import Sock
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import (
@@ -25,14 +23,17 @@ from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers import issue_registry as ir
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .compat import (
+from .const import (
+    CONF_OWLET_EXPIRY,
+    CONF_OWLET_REFRESH,
+    DOMAIN,
     OWLET_CREDENTIAL_ERRORS,
-    OwletConnectionError,
-    OwletDevicesError,
-    OwletError,
+    SUPPORTED_VERSIONS,
 )
-from .const import CONF_OWLET_EXPIRY, CONF_OWLET_REFRESH, DOMAIN, SUPPORTED_VERSIONS
 from .coordinator import OwletCoordinator
+from .owletapi.api import OwletAPI
+from .owletapi.exceptions import OwletConnectionError, OwletDevicesError, OwletError
+from .owletapi.sock import Sock
 
 PLATFORMS: list[Platform] = [Platform.BINARY_SENSOR, Platform.SENSOR, Platform.SWITCH]